chore(deepseek_v4): drop the scratch notes and consolidate the tests
Gets the branch into PR shape, at a net -2668 lines.

- Deletes the three root markdown scratch files (`TODO.md`,
  `NOTES-ds-v4-inference-preflight.md`, `ISSUES_vllm_dsv4.md`), the
  `### DeepSeek V4` section of the `start-run` skill, and `kl-check.toml`,
  repairing every reference that pointed at them.
- Deletes the three `*_hf.py` test modules. They open with
  `importorskip("transformers.models.deepseek_v4")`, which only exists from
  transformers 5.15 against the pinned 5.6.2, so they never ran.
- Merges `deepseek_v4_temp_helpers.py` into `deepseek_v4_helpers.py`,
  renaming the whole-model config to `_MODEL` and the per-module one to
  `_HC` so the two scales no longer collide silently, and salvages 26 of
  `test_deepseek_v4_temp.py`'s cases into `test_deepseek_v4.py` as 14
  parametrized functions.
- Adds `test_deepseek_v4_cpu.py`: the dequantize tests plus the three
  config tests that touch no CUDA and were gpu-marked only by module
  inheritance, so the CPU job gains three tests instead of losing three.
