--- examples/advanced/deepseek-v4-flash/kl-check.toml
@@ -1,78 +1,0 @@
-# KL-mismatch-table run for the docs/development.md "Adding a New Model" merge bar: 20 steps on
-# math-env, batch_size=64. Read the per-step trainer/vLLM mismatch off the `mismatch_kl/all/mean`
-# metric (console log line "Mismatch KL {:.4f}" and W&B) -- there is no separate script that
-# produces this; transcribe the 20 step values into the PR. All entries must be < 0.015 to merge.
-# Runs on 4 train + 8 infer nodes (8xH200 each).
-max_steps = 20
-seq_len = 2048
-
-[model]
-name = "deepseek-ai/DeepSeek-V4-Flash-0731"
-
-[weight_broadcast]
-# NCCL/NIXL broadcast need convert_layer_to_vllm_kernel, which this port does not implement
-# (see TODO.md). Filesystem broadcast uses convert_to_hf instead, which is implemented and tested.
-type = "filesystem"
-
-[deployment]
-type = "multi_node"
-num_train_nodes = 4
-num_infer_nodes = 8
-
-[slurm]
-job_name = "deepseek-v4-flash-kl-check"
-partition = "" # FILL IN
-
-[ckpt]
-
-[monitors.wandb]
-project = "deepseek-v4-flash"
-name = "deepseek-v4-flash-kl-check"
-
-[trainer.model]
-impl = "custom"
-ep = 8
-
-[trainer.model.compile]
-
-[trainer.model.ac]
-freq = 1
-
-[trainer.model.ac_offloading]
-max_inflight_activations = 5
-
-[trainer.optim]
-lr = 3e-6
-
-[orchestrator]
-batch_size = 64
-
-[orchestrator.train.sampling]
-max_completion_tokens = 128
-
-[[orchestrator.train.source]]
-name = "hendrycks-math"
-env.taskset.id = "math-env"
-env.agent.harness.id = "null"
-env.agent.runtime.type = "subprocess"
-
-[inference]
-# Mandatory for DeepSeek V4. See TODO.md.
-use_deep_gemm = true
-
-[inference.vllm]
-tensor_parallel_size = 8
-# One full-TP replica per inference node (8 nodes x TP=8 = 64 GPUs total), required by the
-# multi-node expert-parallel inference validator when enable_expert_parallel is set.
-data_parallel_size = 8
-enable_expert_parallel = true
-# Both required by DeepSeek V4's sparse-MLA KV layout, not tuning choices. See the comments in
-# inference.toml for the assertion these avoid.
-kv_cache_dtype = "fp8"
-block_size = 256
-
-# The checkpoint ships no Jinja chat template, so `name = "default"` resolves to DefaultRenderer
-# and raises "tokenizer.chat_template is not set" on the first rollout. `deepseek-v4` is what
-# auto-resolution would pick anyway; pinned so the choice survives a MODEL_RENDERER_MAP change.
-[orchestrator.renderer]
-name = "deepseek-v4"
